refactor: 테이블별로 repository 분리
